Added just enough tests to get all code analyzed
Coverage testing wasn't analyzing files that weren't touched at all. All files
are now included with the exception of the Database files that I'll be
consolidating soon into a single class (no reason to have so many classes
touched just to create a single database object). Aside from that, all classes
are now represented in the coverage report, Sad to go from 80% down to 50% but
oh well, it's for the best!