Fixed issue with UTF-8 2-byte characters
Characters were being split up and causing invalid sequences when using
`substr()`. Went ahead and updated to use `mb_strcut()` and forcing the
character encoding to UTF-8. I think the plan down the road will be to set the
internal encoding to UTF-8 but I am not currently sure how that could effect
the rest of the system (perhaps it won't). Closes #39