Dropped asset minification
Been moving away from using this functionality in favor of either Grunt or
Gulp. Dropped functionality for minification of CSS and JS but left the dynamic
reference logic intact. This will be moved to the HTML class eventually and
expanded to support the generation of the HTML tags as well as injecting the
URI with a timestamp.