Moved validate logic into resource
I couldn't find a single scenario where I was using the Validate class in my
site level code. Dropped the overhead of calling a static class method multiple
times on a page by moving the logic in the Resource class. Also changed the
response to always return the errors as parameter => array. This allows a
developer to choose if they want to display one error or all of the errors.