Working on basic auth and dropped log class
I personally haven't used it in years so I'm just getting to the mindset that
it's not a very useful piece of functionality. Nginx can easily log all of your
requests and any time I need to troubleshoot SQL (not all that often) I do it
directly without using the class.