Dropped pickles.ph made some assumptions
Wanted to be 100% reliant on the autoloader so I dropped the manual include of
pickles.php. Also moved some of the PHP ini logic out of here and baked in
some assumptions about environments and settings. Basically anything other than
"production" will set display errors to true, production is false.